+++ drivers/staging/android/lowmemorykiller.c
@@ -59,18 +59,22 @@
 static uint32_t lowmem_debug_level = 1;
 static short lowmem_adj[6] = {
 	0,
-	1,
-	6,
-	12,
+	100,
+	200,
+	300,
+	528,
+	900,
 };
-static int lowmem_adj_size = 4;
+static int lowmem_adj_size = 6;
 static int lowmem_minfree[6] = {
-	3 * 512,	/* 6MB */
-	2 * 1024,	/* 8MB */
-	4 * 1024,	/* 16MB */
-	16 * 1024,	/* 64MB */
+	7168,	/* 28 MB */
+	8192,	/* 32 MB */
+	11776,	/* 46 MB */
+	15360,	/* 60 MB */
+	18944,	/* 74 MB */
+	22528,	/* 88 MB */
 };
-static int lowmem_minfree_size = 4;
+static int lowmem_minfree_size = 6;
 static int lmk_fast_run = 1;
 
 static unsigned long lowmem_deathpending_timeout;